feat: Adapt packaging to new folders hierachie
The package is no longer installed statically via the setup.py file. The
new pure Python package is installed in its own folder via the
packaging. This includes the executable files, which are made available
via links in /usr/bin. Here we use the old name 'revpipyload' and not
'revpipyloadd'.

Signed-off-by: Sven Sager <s.sager@kunbus.com>